move failure-analyzer to util package for better reuse
(cherry picked from commit 7ad83c27ef52fee65b2b78e317f2be6550473470)
