+++ core/cas-server-core-util-api/src/main/java/org/apereo/cas/util/spring/boot/BeanDefinitionStoreFailureAnalyzer.java
@@ -1,4 +1,4 @@
-package org.apereo.cas.web;
+package org.apereo.cas.util.spring.boot;
 
 import org.apereo.cas.util.LoggingUtils;
 
@@ -15,14 +15,14 @@ import java.io.PrintWriter;
 import java.io.StringWriter;
 
 /**
- * Failure analyzer for spring boot startup exceptions from BeanDefinitionStoreException.
+ * Failure analyzer for spring boot startup exceptions from {@link BeanDefinitionStoreException}.
  * @author Hal Deadman
  * @since 6.4.0
  */
 @Slf4j
 public class BeanDefinitionStoreFailureAnalyzer extends AbstractFailureAnalyzer<BeanDefinitionStoreException> {
 
-    private static String ANALYSIS = "Review the properties available for the configuration. Enable debug logging on "
+    private static final String ANALYSIS = "Review the properties available for the configuration. Enable debug logging on "
             + BeanDefinitionStoreFailureAnalyzer.class.getName() + " to see exception stack trace";
 
     @Override
@@ -33,7 +33,7 @@ public class BeanDefinitionStoreFailureAnalyzer extends AbstractFailureAnalyzer<
         return new FailureAnalysis(getDescription(cause), ANALYSIS, cause);
     }
 
-    private String getDescription(final BeanDefinitionStoreException ex) {
+    private static String getDescription(final BeanDefinitionStoreException ex) {
         val causedMsg = ExceptionUtils.getRootCauseMessage(ex);
         val description = new StringWriter();
         val printer = new PrintWriter(description);
