+++ core/cas-server-core-util-api/src/test/java/org/apereo/cas/AllUtilTestsSuite.java
@@ -23,6 +23,7 @@ import org.apereo.cas.util.serialization.SerializationUtilsTests;
 import org.apereo.cas.util.spring.ConvertersTests;
 import org.apereo.cas.util.spring.SpringAwareMessageMessageInterpolatorTests;
 import org.apereo.cas.util.spring.SpringExpressionLanguageValueResolverTests;
+import org.apereo.cas.util.spring.boot.BeanDefinitionStoreFailureAnalyzerTests;
 import org.apereo.cas.util.spring.boot.DefaultCasBannerTests;
 import org.apereo.cas.util.ssl.CompositeX509KeyManagerTests;
 import org.apereo.cas.util.ssl.CompositeX509TrustManagerTests;
@@ -55,6 +56,7 @@ import org.junit.runner.RunWith;
     SpringAwareMessageMessageInterpolatorTests.class,
     HttpUtilsTests.class,
     ConvertersTests.class,
+    BeanDefinitionStoreFailureAnalyzerTests.class,
     SimpleHttpClientFactoryBeanTests.class,
     GroovyScriptResourceCacheManagerTests.class,
     LoggingUtilsTests.class,
