+++ core/cas-server-core-util/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.apereo.cas.config.CasCoreUtilConfiguration,\
   org.apereo.cas.config.CasCoreUtilSerializationConfiguration
+
+org.springframework.boot.diagnostics.FailureAnalyzer=org.apereo.cas.util.spring.boot.BeanDefinitionStoreFailureAnalyzer
