--- webapp/cas-server-webapp-init/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.diagnostics.FailureAnalyzer=org.apereo.cas.web.BeanDefinitionStoreFailureAnalyzer
