+++ webapp/cas-server-webapp-init/src/test/java/org/apereo/cas/AllCasWebApplicationTestsSuite.java
@@ -10,10 +10,7 @@ import org.junit.runner.RunWith;
  * @author Misagh Moayyed
  * @since 6.1.0
  */
-@SelectClasses({
-    CasEmbeddedContainerUtilsTests.class,
-    BeanDefinitionStoreFailureAnalyzerTests.class
-})
+@SelectClasses(CasEmbeddedContainerUtilsTests.class)
 @RunWith(JUnitPlatform.class)
 public class AllCasWebApplicationTestsSuite {
 }
